config/requirements/base.txt
django==3.2.0
djangorestframework==3.12.4
environs==9.3.2
psycopg2-binary==2.8.6
ipython==7.22.0
django-extensions==3.1.3
pillow==8.2.0
djangorestframework-simplejwt==4.7.1